Add `last_address` param to `/metadata`
Skip entries that have the same timestamp as `start_timestamp` but where the pubkey hash is less than or equal to the given address' pubkey hash.

This allows specifying the largest timestamp returned from a previous call as `start_timestamp` with the corresponding address, to make sure all entries are retrieved once without duplicates.