Add support for updated votes on a topic
Currently, there's no notification on a topic that a vote occured on an
existing message. This commit adds support for tracking those state
updates in the topic_cf. That will enable Stamp to properly update users
who have already downloaded messages as to the fact that a post's votes
were updated.