[metadata] Move pubkey commitment out of burn transaction
Currently, for burn transactions are expected to be a hash of the
combination of the pubkey and the payload hash. However, this has
consequences for the topic posts which need to have a commitment
to their hash directly on the blockchain, and voting transactions
do not necessarily need to sign the payloads.

On the other hand, for address metadata if the serialized payload does
not commit to the pubkey then the burn transactions could be reusable
to input many addresses into the registry server.

This commit does two things:

1. Repeat the pubkey in the AddressMetadata so the hash of the serialized
   protobuf will commit to the pubkey.
2. Remove the need to commit to the pubkey in the OP_RETURN commitment that
   burns some Lotus.